Sirius input parameters transfered through the interface
SIRIUS parameters can not be given using the cp2k input file standard.
The options are automatically retrieved from the sirius interaface and
are updated eventually with each new version of sirius. If new options
appear in the SIRIUS they are automatically included in cp2k after recompilation.

Generating the documentation should be done the psmp version of cp2k though